refs #13923 Rename builders and updated specs
