+++ app/notices/shop_notices/renewal_employer_eligibility_notice.rb
@@ -1,0 +1,10 @@
+class ShopNotices::RenewalEmployerEligibilityNotice < ShopNotice
+
+  attr_accessor :employer_profile
+
+  def deliver
+    build
+    super
+  end
+
+end
