+++ db/seedfiles/notice_triggers_seed.rb
@@ -73,7 +73,7 @@ shop_notice_triggers = [
       {
         name: 'PlanYear Renewal',
         notice_template: 'notices/shop_notices/3a_employer_plan_year_renewal',
-        notice_builder: 'ShopNotices::EmployerRenewalNotice',
+        notice_builder: 'ShopNotices::RenewalEmployerEligibilityNotice',
         mpi_indicator: 'MPI_SHOPRA',
         notice_trigger_element_group: {
           market_places: ['shop'],
@@ -94,7 +94,7 @@ shop_notice_triggers = [
       {
         name: 'PlanYear Renewal Auto-Published',
         notice_template: 'notices/shop_notices/3b_employer_plan_year_renewal',
-        notice_builder: 'ShopNotices::EmployerRenewalNotice',
+        notice_builder: 'ShopNotices::RenewalEmployerEligibilityNotice',
         mpi_indicator: 'MPI_SHOPRB',
         notice_trigger_element_group: {
           market_places: ['shop'],
@@ -116,7 +116,7 @@ shop_notice_triggers = [
       {
         name: 'Group Renewal Notice',
         notice_template: 'notices/shop_notices/5_employer_renewal_notice',
-        notice_builder: 'ShopNotices::RenewalGroupNotice',
+        notice_builder: 'ShopNotices::EmployerRenewalNotice',
         mpi_indicator: 'MPI_SHOP5',
         notice_trigger_element_group: {
           market_places: ['shop'],
@@ -138,7 +138,7 @@ shop_notice_triggers = [
       {
         name: 'Conversion, Group Renewal Available',
         notice_template: 'notices/shop_notices/6_conversion_group_renewal_notice',
-        notice_builder: 'ShopNotices::ConversionRenewalNotice',
+        notice_builder: 'ShopNotices::EmployerRenewalNotice',
         mpi_indicator: 'MPI_SHOP6',
         notice_trigger_element_group: {
           market_places: ['shop'],
