+++ spec/notices/employer_renewal_notice_spec.rb
@@ -1,6 +1,6 @@
 require 'rails_helper'
 
-RSpec.describe ShopNotices::RenewalGroupNotice do
+RSpec.describe ShopNotices::EmployerRenewalNotice do
   let(:start_on) { TimeKeeper.date_of_record.beginning_of_month + 1.month - 1.year}
   let!(:employer_profile){ create :employer_profile}
   let!(:person){ create :person}
@@ -9,11 +9,11 @@ RSpec.describe ShopNotices::RenewalGroupNotice do
   let!(:renewal_plan_year) { FactoryGirl.create(:plan_year, employer_profile: employer_profile, start_on: start_on + 1.year, :aasm_state => 'renewing_draft' ) }
   let!(:renewal_benefit_group) { FactoryGirl.create(:benefit_group, plan_year: renewal_plan_year, title: "Benefits #{renewal_plan_year.start_on.year}") }
   let(:application_event){ double("ApplicationEventKind",{
-                            :name =>'Group Renewal Notice',
-                            :notice_template => 'notices/shop_notices/5_employer_renewal_notice',
-                            :notice_builder => 'ShopNotices::RenewalGroupNotice',
-                            :mpi_indicator => 'MPI_SHOP5',
-                            :title => "Group Renewal Available"})
+                            :name =>'Conversion, Group Renewal Available',
+                            :notice_template => 'notices/shop_notices/6_conversion_group_renewal_notice',
+                            :notice_builder => 'ShopNotices::EmployerRenewalNotice',
+                            :mpi_indicator => 'MPI_SHOP6',
+                            :title => "Welcome to DC Health Link, Group Renewal Available"})
                           }
     let(:valid_parmas) {{
         :subject => application_event.title,
@@ -27,7 +27,7 @@ RSpec.describe ShopNotices::RenewalGroupNotice do
     end
     context "valid params" do
       it "should initialze" do
-        expect{ShopNotices::RenewalGroupNotice.new(employer_profile, valid_parmas)}.not_to raise_error
+        expect{ShopNotices::EmployerRenewalNotice.new(employer_profile, valid_parmas)}.not_to raise_error
       end
     end
 
@@ -35,7 +35,7 @@ RSpec.describe ShopNotices::RenewalGroupNotice do
       [:mpi_indicator,:subject,:template].each do  |key|
         it "should NOT initialze with out #{key}" do
           valid_parmas.delete(key)
-          expect{ShopNotices::RenewalGroupNotice.new(employer_profile, valid_parmas)}.to raise_error(RuntimeError,"Required params #{key} not present")
+          expect{ShopNotices::EmployerRenewalNotice.new(employer_profile, valid_parmas)}.to raise_error(RuntimeError,"Required params #{key} not present")
         end
       end
     end
@@ -44,7 +44,7 @@ RSpec.describe ShopNotices::RenewalGroupNotice do
   describe "Build" do
     before do
       allow(employer_profile).to receive_message_chain("staff_roles.first").and_return(person)
-      @employer_notice = ShopNotices::RenewalGroupNotice.new(employer_profile, valid_parmas)
+      @employer_notice = ShopNotices::EmployerRenewalNotice.new(employer_profile, valid_parmas)
     end
     it "should build notice with all necessory info" do
       @employer_notice.build
@@ -57,13 +57,14 @@ RSpec.describe ShopNotices::RenewalGroupNotice do
   describe "append data" do
     before do
       allow(employer_profile).to receive_message_chain("staff_roles.first").and_return(person)
-      @employer_notice = ShopNotices::RenewalGroupNotice.new(employer_profile, valid_parmas)
+      @employer_notice = ShopNotices::EmployerRenewalNotice.new(employer_profile, valid_parmas)
     end
     it "should append data" do
       renewing_plan_year = employer_profile.plan_years.where(:aasm_state => "renewing_draft").first
       @employer_notice.append_data
-      expect(@employer_notice.notice.plan_year.open_enrollment_end_on).to eq renewing_plan_year.open_enrollment_end_on
       expect(@employer_notice.notice.plan_year.start_on).to eq renewing_plan_year.start_on
+      expect(@employer_notice.notice.plan_year.open_enrollment_end_on).to eq renewing_plan_year.open_enrollment_end_on
+      expect(@employer_notice.notice.plan_year.carrier_name).to eq renewing_plan_year.benefit_groups.first.reference_plan.carrier_profile.legal_name
     end
   end
